test(api): migrate ApiResponse formatting tests to integration test class
